+++ src/PhpSpreadsheet/Shared/OLERead.php
@@ -326,14 +326,6 @@ class OLERead
      */
     private static function getInt4d($data, $pos)
     {
-        if (trim($data) == '') {
-            // No data provided
-            throw new ReaderException('Parameter data is empty.');
-        } elseif ($pos < 0) {
-            // Invalid position
-            throw new ReaderException('Parameter pos=' . $pos . ' is invalid.');
-        }
-
         $len = strlen($data);
         if ($len < $pos + 4) {
             $data .= str_repeat("\0", $pos + 4 - $len);